tests: load fixtures only if there are some
git-svn-id: file:///var/www/production/sfweb/svn/trunk@2163 ee427ae8-e902-0410-961c-c3ed070cd9f9

+++ test/functionnal/bootstrap.php
@@ -19,24 +19,30 @@ $sf_symfony_data_dir = realpath(dirname(__FILE__).'/../../data');
 // initialize symfony
 require_once(SF_ROOT_DIR.DIRECTORY_SEPARATOR.'apps'.DIRECTORY_SEPARATOR.SF_APP.DIRECTORY_SEPARATOR.'config'.DIRECTORY_SEPARATOR.'config.php');
 
-// initialize database manager
-$databaseManager = new sfDatabaseManager();
-$databaseManager->initialize();
+// remove all cache
+sfToolkit::clearDirectory(sfConfig::get('sf_cache_dir'));
 
-// cleanup database
-$db = sfConfig::get('sf_data_dir').DIRECTORY_SEPARATOR.'/database.sqlite';
-if (file_exists($db))
+if (isset($fixtures))
 {
-  unlink($db);
+  // initialize database manager
+  $databaseManager = new sfDatabaseManager();
+  $databaseManager->initialize();
+
+  // cleanup database
+  $db = sfConfig::get('sf_data_dir').DIRECTORY_SEPARATOR.'/database.sqlite';
+  if (file_exists($db))
+  {
+    unlink($db);
+  }
+
+  // initialize database
+  $sql = file_get_contents(sfConfig::get('sf_data_dir').DIRECTORY_SEPARATOR.'sql'.DIRECTORY_SEPARATOR.'lib.model.schema.sql');
+  $sql = preg_replace('/^\s*\-\-.+$/m', '', $sql);
+  $sql = preg_replace('/^\s*DROP TABLE .+?$/m', '', $sql);
+  $con = Propel::getConnection();
+  $con->executeQuery($sql);
+
+  // load fixtures
+  $data = new sfPropelData();
+  $data->loadDataFromArray($fixtures);
 }
-
-// initialize database
-$sql = file_get_contents(sfConfig::get('sf_data_dir').DIRECTORY_SEPARATOR.'sql'.DIRECTORY_SEPARATOR.'lib.model.schema.sql');
-$sql = preg_replace('/^\s*\-\-.+$/m', '', $sql);
-$sql = preg_replace('/^\s*DROP TABLE .+?$/m', '', $sql);
-$con = Propel::getConnection();
-$con->executeQuery($sql);
-
-// load fixtures
-$data = new sfPropelData();
-$data->loadDataFromArray($fixtures);
